+++ backend/config/supabaseDB.js
@@ -12,3 +12,4 @@ pool.connect()
 
 module.exports = pool;
 
+