+++ backend/models/index.js
@@ -1,0 +1,7 @@
+module.exports = {
+  ExamPackage: require('./examPackage.model'),
+  AiAuditTrail: require('./aiAuditTrail.model'),
+  ProctoringEvent: require('./proctoringEvent.model'),
+  Incident: require('./incident.model'),
+};
+