+++ backend/routes/admin.js
@@ -1,9 +1,0 @@
-const router = require('express').Router();
-const postCourseOverride = require('../controllers/postCourseOverride');
-const { requireScope } = require('../utils/auth');
-
-router.post('/override/post-course', requireScope('manage:assessments'), postCourseOverride);
-
-module.exports = router;
-
-